// qto-ui/src/app/models/activation-attempt.model.ts
import { AbstractBaseModel } from "./abstract-base-model";
import { ActivationAttemptRequirement } from "./activation-attempt-requirement.model";

export class ActivationAttempt extends AbstractBaseModel {
  scheduledAttemptStatus: string;
  internalTechAssigned: string;
  fieldDispatchVendor: string;
  fieldTechName: string;
  fieldTechPhone: string;
  fieldTechCheckIn: Date | null;
  fieldTechCheckOut: Date | null;
  testedDownloadSpeed: string;
  testedUploadSpeed: string;
  latency: string;
  backupDownloadSpeed: string;
  backupUploadSpeed: string;
  signalRsrp: string;
  sinrRsrq: string;
  locationDowntownForCutover: string;
  closeoutCode: string;
  networkCompleteDate: Date | null;
  voipCompleteDate: Date | null;
  issueNotes: string;
  issueNotesDisplay: string;
  closeNotes: string;
  duplicateToRelated: boolean;
  poNumber: string;
  serviceId: number;
  scheduledCheckInTime: number;
  activationScheduleId: number;
  attemptNumber: number;
  requirements: ActivationAttemptRequirement[];
  replace4g5g: number;
  primaryUid: string;
  secondaryUid: string;
  ftdiVendorId: string;
  ftdiVendorStatus: string;
  managedRouterSerialNumber: string;
  cancelledDate: Date;
  cancelledBy: string;
  sameDaySchedule: boolean;
  useExistingInventoryLocationAddress: boolean;
}